move eventual check
