+++ crates/core/src/lib.rs
@@ -11,6 +11,7 @@ pub mod responsechannel;
 pub mod signal;
 pub mod strng;
 pub mod telemetry;
+pub mod test_helpers;
 pub mod timestamp;
 mod tokio_metrics;
 pub mod version;
