+++ crates/core/src/telemetry.rs
@@ -587,8 +587,7 @@ pub mod testing {
 	use std::io;
 	use std::io::IoSlice;
 	use std::sync::{Arc, Mutex, MutexGuard, OnceLock};
-	use std::time::{Duration, SystemTime};
-	use tracing::trace;
+	use std::time::Duration;
 	use tracing_subscriber::fmt;
 	use tracing_subscriber::fmt::writer::Tee;
 	use tracing_subscriber::layer::SubscriberExt;
@@ -624,39 +623,10 @@ pub mod testing {
 		}
 	}
 
-	/// check_eventually runs a function many times until it reaches the expected result.
-	/// If it doesn't the last result is returned
-	async fn check_eventually<F, CF, T, Fut>(dur: Duration, f: F, expected: CF) -> Result<T, T>
-	where
-		F: Fn() -> Fut,
-		Fut: Future<Output = T>,
-		T: Eq + Debug,
-		CF: Fn(&T) -> bool,
-	{
-		use std::ops::Add;
-		let mut delay = Duration::from_millis(10);
-		let end = SystemTime::now().add(dur);
-		let mut last: T;
-		let mut attempts = 0;
-		loop {
-			attempts += 1;
-			last = f().await;
-			if expected(&last) {
-				return Ok(last);
-			}
-			trace!("attempt {attempts} with delay {delay:?}");
-			if SystemTime::now().add(delay) > end {
-				return Err(last);
-			}
-			tokio::time::sleep(delay).await;
-			delay *= 2;
-		}
-	}
-
 	/// eventually_find waits until at least one log line matches the given keys.
 	/// If not found, panics
 	pub async fn eventually_find(want: &[(&str, &str)]) -> Option<Value> {
-		check_eventually(
+		crate::test_helpers::check_eventually(
 			Duration::from_secs(1),
 			|| async { find(want).into_iter().next() },
 			|log| log.is_some(),
